rename: useGetNoticeDetail 폴더 이동
